Recibo térmico: dotDensity single en vez de double, no ocupaba el ancho del papel
'double' achicaba la imagen impresa dejando una franja en blanco a la
derecha del papel (confirmado con foto real). Mismo síntoma ya
encontrado y corregido antes en el ticket adhesivo (misma marca SAT,
impresora distinta) — 'single' sí ocupa el ancho físico completo.

+++ app-sastreria/src/app/core/services/qz-print.service.ts
@@ -231,7 +231,11 @@ export class QzPrintService {
         type: 'raw',
         format: 'image',
         data: dataUrl,
-        options: { language: 'ESCPOS', dotDensity: 'double', quantization: 'black', threshold: 128 },
+        // dotDensity:'double' achicaba la imagen dentro del ancho físico del
+        // papel (dejaba una franja en blanco a la derecha) — mismo síntoma ya
+        // encontrado y corregido en el ticket adhesivo (misma marca SAT, otro
+        // equipo), donde 'single' sí ocupó el ancho completo.
+        options: { language: 'ESCPOS', dotDensity: 'single', quantization: 'black', threshold: 128 },
       },
       {
         type: 'raw',
